+++ apps/api/internal/db/queries/polling_unit_assignments.sql.go
@@ -28,7 +28,7 @@ type CreateAssignmentParams struct {
 	UserID          int64       `json:"user_id"`
 	PollingUnitID   int32       `json:"polling_unit_id"`
 	ElectionGroupID int64       `json:"election_group_id"`
-	PartyID         int64       `json:"party_id"`
+	PartyID         int16       `json:"party_id"`
 	RoleType        pgtype.Text `json:"role_type"`
 	AssignedBy      pgtype.Int8 `json:"assigned_by"`
 }
@@ -121,7 +121,7 @@ type GetAssignmentByIDRow struct {
 	UserID                  int64              `json:"user_id"`
 	PollingUnitID           int32              `json:"polling_unit_id"`
 	ElectionGroupID         int64              `json:"election_group_id"`
-	PartyID                 int64              `json:"party_id"`
+	PartyID                 int16              `json:"party_id"`
 	RoleType                pgtype.Text        `json:"role_type"`
 	AssignedBy              pgtype.Int8        `json:"assigned_by"`
 	ArrivedAt               pgtype.Timestamptz `json:"arrived_at"`
@@ -198,7 +198,7 @@ ORDER BY gs.threshold
 `
 
 type GetPartyElectionGroupCoverageDistributionParams struct {
-	PartyID         int64 `json:"party_id"`
+	PartyID         int16 `json:"party_id"`
 	ElectionGroupID int64 `json:"election_group_id"`
 }
 
@@ -284,7 +284,7 @@ type ListAssignmentsRow struct {
 	UserID                  int64              `json:"user_id"`
 	PollingUnitID           int32              `json:"polling_unit_id"`
 	ElectionGroupID         int64              `json:"election_group_id"`
-	PartyID                 int64              `json:"party_id"`
+	PartyID                 int16              `json:"party_id"`
 	RoleType                pgtype.Text        `json:"role_type"`
 	AssignedBy              pgtype.Int8        `json:"assigned_by"`
 	CreatedAt               pgtype.Timestamptz `json:"created_at"`
@@ -395,7 +395,7 @@ type UpdateAssignmentTrackingRow struct {
 	UserID                  int64              `json:"user_id"`
 	PollingUnitID           int32              `json:"polling_unit_id"`
 	ElectionGroupID         int64              `json:"election_group_id"`
-	PartyID                 int64              `json:"party_id"`
+	PartyID                 int16              `json:"party_id"`
 	RoleType                pgtype.Text        `json:"role_type"`
 	AssignedBy              pgtype.Int8        `json:"assigned_by"`
 	ArrivedAt               pgtype.Timestamptz `json:"arrived_at"`
