+++ apps/api/internal/utils/jwt.go
@@ -18,7 +18,7 @@ type JWTClaims struct {
 	FakeID    int64    `json:"fake_id"`
 	Username  string   `json:"username"`
 	Roles     []string `json:"roles"`
-	PartyID   int64    `json:"party_id,omitempty"`
+	PartyID   int16    `json:"party_id,omitempty"`
 	jwt.RegisteredClaims
 }
 
@@ -33,8 +33,8 @@ func (c *JWTClaims) HasRole(role string) bool {
 }
 
 // GenerateToken creates a signed JWT with the given claims, secret, and duration.
-func GenerateToken(userID int64, fakeID int64, username string, roles []string, secret string, duration time.Duration, partyID ...int64) (string, error) {
-	var pid int64
+func GenerateToken(userID int64, fakeID int64, username string, roles []string, secret string, duration time.Duration, partyID ...int16) (string, error) {
+	var pid int16
 	if len(partyID) > 0 {
 		pid = partyID[0]
 	}
